+++ src/classes/babies/BurningBaby.ts
@@ -1,0 +1,16 @@
+import { CacheFlag, ModCallback } from "isaac-typescript-definitions";
+import { Callback } from "isaacscript-common";
+import { Baby } from "../Baby";
+
+/** Spawns a Mega Troll Bomb every N seconds. */
+export class BurningBaby extends Baby {
+  override onAdd(player: EntityPlayer): void {
+    player.AddCacheFlags(CacheFlag.LUCK);
+    player.EvaluateItems();
+  }
+
+  @Callback(ModCallback.EVALUATE_CACHE, CacheFlag.LUCK)
+  evaluateCacheLuck(player: EntityPlayer): void {
+    player.Luck += 40;
+  }
+}