+++ docs/history.md
@@ -10,6 +10,7 @@
   - Lowface Baby (#73) - Starts with Book of Virtues + Unicorn Stump
   - Eyebrows Baby (#89) - Starts with 5x Guppy's Hair Ball + flight + blindfolded
   - Lipstick Baby (#105) - Range oscillates
+  - Burning Baby (#467) - Starts with Vasculitis + A Lighter + 40 Luck
   - Tooth Baby (#478) - Starts with Black Tooth
   - Thirteenth Baby (#533) - Starts with Modeling Clay
   - Robo-Baby (#578) - Starts with Technology + 3x Robo-Baby